requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0          # Parallel execution
pytest-timeout==2.2.0        # Test timeouts
pytest-dependency==0.5.1     # Test dependencies

# Allure Reporting
allure-pytest==2.13.2

# Hardware Interface Libraries
python-can==4.3.1            # CAN bus support (SocketCAN, Vector, PCAN)
pyserial==3.5                # Serial/UART communication
can-isotp==2.0.7             # ISO-TP for diagnostics

# UDS/Diagnostics (optional - install if needed)
# udsoncan==1.20.1
# python-doip==1.0.5

# Configuration Management
PyYAML==6.0.1
pydantic==2.5.3              # Config validation
python-dotenv==1.0.0         # Environment variables

# Utilities
colorlog==6.8.0              # Colored logging
tabulate==0.9.0              # Pretty tables
click==8.1.7                 # CLI framework
